refactor: move response models
