+++ internal/api/http/handler/v1/v1.go
@@ -1,1 +1,16 @@
 package v1
+
+import (
+	ierr "github.com/gxravel/bus-routes-visualizer/internal/errors"
+)
+
+type RangeItemsResponse struct {
+	Items interface{} `json:"items"`
+	Total int64       `json:"total"`
+}
+
+// Response describes http response for api v1.
+type Response struct {
+	Data  interface{}    `json:"data,omitempty"`
+	Error *ierr.APIError `json:"error,omitempty"`
+}
